pyproject.toml
[project]
name = "ai4gd-momconnect-haystack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "weaviate-haystack>=6.0.0",
    "haystack-ai>=2.13.2",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "fastapi[standard]>=0.115.12",
    "sentry-sdk[fastapi]>=2.29.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "sqlalchemy>=2.0.29",
    "aiosqlite>=0.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.3.5",
    "ruff>=0.11.12",
    "deepeval>=0.24.38",
]

[project.scripts]
main = "ai4gd_momconnect_haystack:main.main"

[tool.uv]
package = true

[[tool.mypy.overrides]]
module = "haystack.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "haystack_integrations.*"
ignore_missing_imports = true
